+++ crates/affinidi-did-resolver/affinidi-did-resolver-methods/did-webvh/src/lib.rs
@@ -43,6 +43,8 @@ pub enum DIDWebVHError {
     ValidationError(String),
     #[error("DeactivatedError: {0}")]
     DeactivatedError(String),
+    #[error("DIDError: {0}")]
+    DIDError(String),
 }
 
 pub struct DIDWebVH;